don't cast 0 to (long)

+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/windmill/WindmillServerBase.java
@@ -112,7 +112,7 @@ public class WindmillServerBase extends WindmillServerStub {
   @Override
   public long getAndResetThrottleTime() {
     // Windmill appliance does not have throttling.
-    return (long) 0;
+    return 0;
   }
 
   @Override
